+++ sdk/core/azure-core/CHANGELOG.md
@@ -16,8 +16,7 @@
   - `GetMinorVersion`.
 - Changes to `Azure::Core::Context`:
   - Renamed `Get()` to `GetValue()`.
-  - Renamed `WithValue()` to `CreateChildContext()`.
-  - Renamed `WithDeadline()` to `CreateChildContext()`, and changed input parameter type to `Azure::DateTime`.
+  - Changed input parameter type of `WithDeadline()` to `Azure::DateTime`.
 - Removed from `Azure::Core::Http::Policies` namespace: `HttpPolicyOrder`, `TransportPolicy`, `RetryPolicy`, `RequestIdPolicy`, `TelemetryPolicy`, `BearerTokenAuthenticationPolicy`, `LogPolicy`.
 - Renamed `Azure::Core::Http::RawResponse::GetBodyStream()` to `ExtractBodyStream()`.
 - Introduced `Azure::Core::Context::Key` class which takes place of `std::string` used for `Azure::Core::Context` keys previously.